+++ src/types/goal.ts
@@ -2,4 +2,5 @@ export type GoalSettings = {
   currentNetWorth: number
   targetNetWorth: number
   monthlyContribution: number
+  maxExposure: number
 }